Moved read_file/write_file methods for classes other than GameFile from OGFILE2 to OGFILE_OTHER.
Instead of a static, the loaded random seed is now a member of Sys.
